Border util 이동
